+++ CHANGES.md
@@ -8,6 +8,9 @@
     changed from `KeySym` to `(KeyMask, KeySym)`. The default value
     for this is still binded to `Tab` key.
 
+  * New constructor `CenteredAt Rational Rational` added for
+    `XMonad.Prompt.XPPosition`.
+
 ## 0.12 (December 14, 2015)
 
 ### Breaking Changes